Ships now transfer wares

+++ src/components/storage.rs
@@ -2,15 +2,12 @@ use bevy::prelude::Component;
 
 #[derive(Component)]
 pub struct Storage {
-    pub capacity: f32,
-    pub used: f32,
+    pub capacity: u32,
+    pub used: u32,
 }
 
 impl Storage {
-    pub fn new(capacity: f32) -> Self {
-        Self {
-            capacity,
-            used: 0.0,
-        }
+    pub fn new(capacity: u32) -> Self {
+        Self { capacity, used: 0 }
     }
 }
